Edit

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -2,7 +2,7 @@
 -- that divides (and returns) the first by the second number or returns 0
 
 DELIMITER //
-CREATE FUNCTION SafeDiv(a INT, b INT) RETURNS INT
+CREATE FUNCTION SafeDiv(a INT, b INT) RETURNS FLOAT
 BEGIN
   IF b = 0 THEN
     RETURN 0;
